list of known security vulnerabilities
we need to detect when a new vulnerability is being detected by
aqua scanner in the product containers to address the problem and
update the repo if a fix is available.
thus we need a list of already processed cves to evict from result
of a new scan. we store the list in a file inside the
container-builder-src container and use the file contents later on
the scanner side.

partial-jira-bug: CEM-17554
Change-Id: Id00f87623e61094761e264c2a179a3d0c53578d9