Alarms not generating when one controller is down
Right now the default replication factor for kafka topic is 1. So if 1 of the 3
nodes goes down where kafka broker running on that node was the leader then
the messages published on kafka bus is lost and hence alarm-gen is not able to
raise any alarms.
To tolerate up to N-1 server failures we need to set replication factor to N and
min insync replica to N-1.

Closes-Jira-Bug: CEM-17678
Change-Id: I470e90bd0d541fa650161b5cc27217dacbea2ec3