fixes kafka topics wrong replication factor
server.properties file doesn't have a newline in the end thus
the entrypoint appends to the last line. it is not empty thus
default.replication.factor obtain default value 1 on first start
of the container. if controller accesses the kafka server it
generates topics with the factor 1. as the append adds a new
line to the config next start produces correct setup but topics
still use the 1 factor.
let's add an explicit new line to workaround. anyway an empty line
is ok thus we won't break anything if an initial server.properties
ends with a newline.

Change-Id: I33d01daa4d3b86a4a0e86e0c9148ce0e6f1780b0