Update patch set 8
Patch Set 8: Code-Review-2

I'm sorry but I can't approve this change for the following reason.
Initial multiple workers change introduced recently affected user-interface (container parameters) but it contained only one parameter changed, was approved by Michael and was done in accordance with existing mechanism in Openstack. That's why we've bypassed the necessity for full TF blueprint approval and merged the code. 
This particular change:
1. Changes more interfaces and existing behavior.
2. Is not approved by Michael
3. Such mechanism is not present in Openstack
4. It starts introducing orchestration of some internal elements (ports, here) inside previosly straightforward scalar environment parameters. I would personally preferred to avoid that. And there is no guarantee if we go this route, there won't be many more parameters in the future which we would have to start supplying in lists in similar manner for those internal workers.
5. Why only 3 ports are chosen for default? Why not 1 or not 7? I thought the idea was for customer to decide about the scalability required.

On a side note, it's a generic mechanism (to check and assign ports from list to some destination) so I would expect a common function to be used for that when the decision is approved to go that route strategically.

Patch-set: 8
Reviewer: Gerrit User 445 <445@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Label: Code-Review=-2