let's the first server form the cluster
this is required to workaround the unpredictable order the servers
start plus connectivity issues leading to start of independent
clusters.
the list of servers is sorted by its ips thus all the servers see
the same order. we select the first server to form the cluster
and force the rest to wait for its availability, then they join
the cluster.

Change-Id: I79978222778d99e346689daf7ffa18488202a845
Signed-off-by: sgalas@progmaticlab.com