try to perform graceful cassandra process stop
there is a handle stopdaemon recomended to execute before killing
of a cassandra java process, they say it flushes in memory data
on disk and etc. it is required to prevent cassandra data corruption
which annoys in ha mode. let's call the handle from a TERM signal
handler inside the entrypoint script, then kill the java process.

Change-Id: I3065ac83b50294c5250a08267775214f3ff76187
Signed-off-by: sgalas@progmaticlab.com