ansible version change from 2.6.13 to 2.8.2 in device manager container
As part of CEM-25513: Device Manager tmp directories created by fabric-ansible jobs are not cleaned upFIX COMMITTED, we have changed code as per ansible 2.8 version. and still device manager container created with ansible 2.6.13, which is not compatible to 2.8 version change code(playbooks are not exucuted).

As Work Around for this issue is manually we need to upgrade ansible from 2.6.13 to 2.8.2(pip install ansible==2.8.2) in device manage container then only modified code is compatible to execute playbooks.

Tried to change ansible version from 2.6.13 to 2.8.2 in device manager Dockerfile and ran the playbooks, all are executed as expected.

Signed-off-by: Shiva Yaniki <syaniki@juniper.net>
Change-Id: If255d7a7f1982417c6f7033156cd47a225f6f1e2